requirements.txt
Django==2.0.13
django-ckeditor==5.7.1
easy-thumbnails==2.6
